services/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
paho-mqtt==1.6.1
influxdb-client==1.49.0
psycopg[binary]==3.2.10
httpx==0.28.1
scikit-learn==1.7.1
numpy==2.3.3
pandas==2.3.3
pytest==8.4.1
